[1/2] Change TclObject list size/index from 'unsigned' to 'size_t'
In Tcl version 8, the size of a list is represented by the 'int' type
(32-bit). Therefor in our TclObject wrapper class we used 'unsigned'.
But in Tcl version 9 this type was changed the 'Tcl_Size' which is a
typedef for 'ptrdiff_t' (64-bit). So that would cause truncation in our
TclObject wrapper.

As a preparation for that (eventual) change, this patch changes our type
from 'unsigned' (32-bit) to 'size_t' (64-bit). That's OK when using Tcl8
and (later) required when using Tcl9.

I checked all users of TclObject that do something with list sizes or
indices, and adjusted them all. This didn't need to many changes (much
less than I was expecting), also because many users already used 'auto'
to make the code type-agnostic.

+++ src/commands/TclObject.cc
@@ -152,7 +152,7 @@ std::span<const uint8_t> TclObject::getBinary() const
 	return {buf, size_t(length)};
 }
 
-unsigned TclObject::getListLength(Interpreter& interp_) const
+size_t TclObject::getListLength(Interpreter& interp_) const
 {
 	auto* interp = interp_.interp;
 	int result;
@@ -161,7 +161,7 @@ unsigned TclObject::getListLength(Interpreter& interp_) const
 	}
 	return result;
 }
-unsigned TclObject::getListLengthUnchecked() const
+size_t TclObject::getListLengthUnchecked() const
 {
 	int result;
 	if (Tcl_ListObjLength(nullptr, obj, &result) != TCL_OK) {
@@ -170,7 +170,7 @@ unsigned TclObject::getListLengthUnchecked() const
 	return result;
 }
 
-TclObject TclObject::getListIndex(Interpreter& interp_, unsigned index) const
+TclObject TclObject::getListIndex(Interpreter& interp_, size_t index) const
 {
 	auto* interp = interp_.interp;
 	Tcl_Obj* element;
@@ -179,7 +179,7 @@ TclObject TclObject::getListIndex(Interpreter& interp_, unsigned index) const
 	}
 	return element ? TclObject(element) : TclObject();
 }
-TclObject TclObject::getListIndexUnchecked(unsigned index) const
+TclObject TclObject::getListIndexUnchecked(size_t index) const
 {
 	Tcl_Obj* element;
 	if (Tcl_ListObjIndex(nullptr, obj, narrow<int>(index), &element) != TCL_OK) {
@@ -188,7 +188,7 @@ TclObject TclObject::getListIndexUnchecked(unsigned index) const
 	return element ? TclObject(element) : TclObject();
 }
 
-void TclObject::removeListIndex(Interpreter& interp_, unsigned index)
+void TclObject::removeListIndex(Interpreter& interp_, size_t index)
 {
 	unshare(obj);
 	auto* interp = interp_.interp;
